// src/interfaces/Vatsim.ts
export interface vatsimSession {
    "cid": number;
    "name": string;
    "callsign": string;
    "frequency": string;
    "facility": number;
    "rating": number;
    "server": string;
    "visual_range": number;
    "text_atis": string[] | null;
    "last_updated": string;
    "logon_time": string;
}